Added block check to agent xfer channel.
Added a flag to warn us if we're going on block adding a message
to the agents xfer channel. Ran this with a kubectl cp of a 20M
file and had to bump the channel to size 150 for it not to block.
Also added a few related messages.

Factored in feedback from Jefftree and Rata.