konn-client: Stop goroutine if DialResp is not recognized
If the DialResp is not recognized, it seems we never will receive a
close request in this stream. So, we can close this connection by
finishing the goroutine (a defer makes sure the connection is closed
when we finish)

This patch greatly helps to reduce the leaks reported in #276. Although,
this doesn't completely fix the issue.

Updates: #276.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>